+++ CHANGELOG.md
@@ -4,6 +4,7 @@ miniSphere Changelog
 vX.X.X - TBD
 ------------
 
+* Adds `DirectoryStream#dispose()` to the API, for closing an open directory.
 * Renames `SoundStream#buffer()` to `SoundStream#write()`.